+++ server/internal/info/BUILD.bazel
@@ -5,5 +5,8 @@ go_library(
     srcs = ["info.go"],
     importpath = "github.com/pingcap/tidb/server/internal/info",
     visibility = ["//server:__subpackages__"],
-    deps = ["//domain/infosync"],
+    deps = [
+        "//domain/infosync",
+        "//parser/model",
+    ],
 )
